# perf(ext/url): lazy cache for URL keys