--- a/cli/tests/testdata/fetch_response_finalization.js.out
+++ b/cli/tests/testdata/fetch_response_finalization.js.out
@@ -1,2 +1,2 @@
-{ "0": "stdin", "1": "stdout", "2": "stderr", "5": "fetchResponseBody" }
+{ "0": "stdin", "1": "stdout", "2": "stderr", "3": "url", "6": "fetchResponseBody" }
 { "0": "stdin", "1": "stdout", "2": "stderr" }