--- a/ext/url/00_url.js
+++ b/ext/url/00_url.js
@@ -11,6 +11,7 @@
   const core = window.Deno.core;
   const webidl = window.__bootstrap.webidl;
   const {
+    Array,
     ArrayIsArray,
     ArrayPrototypeMap,
     ArrayPrototypePush,