--- a/ext/url/lib.rs
+++ b/ext/url/lib.rs
@@ -14,6 +14,7 @@ use deno_core::Extension;
 use deno_core::OpState;
 use deno_core::Resource;
 use deno_core::ZeroCopyBuf;
+use std::cell::Ref;
 use std::cell::RefCell;
 use std::path::PathBuf;
 
@@ -45,7 +46,7 @@ struct UrlResource {
 
 impl Resource for UrlResource {
   fn name(&self) -> std::borrow::Cow<str> {
-    "url".into()
+    "urlResource".into()
   }
 }