--- a/runtime/js/40_testing.js
+++ b/runtime/js/40_testing.js
@@ -16,6 +16,7 @@
     ArrayPrototypeShift,
     ArrayPrototypeSome,
     ArrayPrototypeSort,
+    ArrayPrototypeIncludes,
     DateNow,
     Error,
     FunctionPrototype,