--- a/runtime/js/40_testing.js
+++ b/runtime/js/40_testing.js
@@ -397,7 +397,7 @@
         const preResource = pre[resource];
         const postResource = post[resource];
         if (preResource === postResource) continue;
-        if (ArrayPrototypeIncludes(resourceAllowList, preResource)) continue;
+        if (ArrayPrototypeIncludes(resourceAllowList, postResource)) continue;
 
         if (preResource === undefined) {
           const [name, action1, action2] = prettyResourceNames(postResource);