This commits adds a key cache for URL parts. This has real performance
impact, like in Fresh -> https://github.com/denoland/fresh/pull/368

```
# This patch
url_parse:              n = 50000, dt = 0.043s, r = 1162791/s, t = 859ns/op
url_parse_get_path:     n = 50000, dt = 0.002s, r = 25000000/s, t = 40ns/op
```

```
# main
url_parse:              n = 50000, dt = 0.066s, r = 757576/s, t = 1320ns/op
url_parse_get_path:     n = 50000, dt = 0.002s, r = 25000000/s, t = 40ns/op
```