@littledivy 2022-09-10
No longer needed. See #15663